requirements/security-audit.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
sqlalchemy>=2.0.39,<3.0.0
psycopg[binary]>=3.2.6,<4.0.0
PyJWT>=2.10.0,<3.0.0
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.10.0,<3.0.0
